requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
streamlit>=1.29.0
pymupdf>=1.24.10
pyyaml>=6.0
slowapi>=0.1.8
python-multipart
pydantic>=2.5.0
cryptography>=41.0.0